move chart components to components/charts
