+++ src/components/charts/YOEBar.jsx
@@ -2,13 +2,13 @@ import { useState } from "react";
 import { ResponsiveBar } from "@nivo/bar";
 
 import Tooltip from "./Tooltip";
-import SocialSharing from "./SocialSharing";
+import SocialSharing from "../SocialSharing";
 
-import theme from "../lib/nivo";
-import { getPercentage } from "../lib/helpers";
-import { TOTAL_RESPONSES } from "../lib/constants";
+import theme from "../../lib/nivo";
+import { getPercentage } from "../../lib/helpers";
+import { TOTAL_RESPONSES } from "../../lib/constants";
 
-import yoeJson from "../data/charts/yoe.json";
+import yoeJson from "../../data/charts/yoe.json";
 const { yoeData, count } = yoeJson;
 
 const YoeKeys = {
